Refactor kodu Sonnet
