+++ custom_components/unifi_gateway_refactored/__init__.py
@@ -1,156 +1,34 @@
 """The UniFi Gateway Dashboard Analyzer integration."""
 from __future__ import annotations
 
-import hashlib
 import logging
 from datetime import timedelta
-from functools import partial
-from collections.abc import Mapping
-from typing import Any, Iterable, Optional, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:  # pragma: no cover - imported for type checking only
     from homeassistant.config_entries import ConfigEntry
     from homeassistant.core import HomeAssistant
-    from homeassistant.helpers import entity_registry as er
     from homeassistant.helpers.typing import ConfigType
+
+__all__ = ["async_setup", "async_setup_entry", "async_unload_entry"]
 
 from .const import (
     CONF_HOST,
-    CONF_PASSWORD,
-    CONF_PORT,
-    CONF_SITE_ID,
-    CONF_TIMEOUT,
-    CONF_USE_PROXY_PREFIX,
-    CONF_USERNAME,
-    CONF_VERIFY_SSL,
     CONF_SPEEDTEST_ENTITIES,
-    CONF_SPEEDTEST_INTERVAL,
-    CONF_WIFI_GUEST,
-    CONF_WIFI_IOT,
     DATA_RUNNER,
     DATA_UNDO_TIMER,
-    DEFAULT_PORT,
-    DEFAULT_SITE,
     DEFAULT_SPEEDTEST_ENTITIES,
-    DEFAULT_SPEEDTEST_INTERVAL,
     DEFAULT_SPEEDTEST_INTERVAL_MINUTES,
-    DEFAULT_TIMEOUT,
-    DEFAULT_USE_PROXY_PREFIX,
-    DEFAULT_VERIFY_SSL,
     DOMAIN,
-    LEGACY_CONF_SPEEDTEST_INTERVAL_MIN,
     PLATFORMS,
 )
-from .coordinator import UniFiGatewayData, UniFiGatewayDataUpdateCoordinator
-from .unifi_client import APIError, AuthError, ConnectivityError, UniFiOSClient
+from .config import normalize_speedtest_entity_ids, resolve_speedtest_interval_seconds, get_wifi_settings
+from .client_factory import async_get_client
+from .coordinator import UniFiGatewayDataUpdateCoordinator
+from .migrations import async_migrate_speedtest_button_unique_id, async_migrate_interface_unique_ids
 from .monitor import SpeedtestRunner
 
 _LOGGER = logging.getLogger(__name__)
-
-
-def _split_entity_candidates(text: str) -> Iterable[str]:
-    for candidate in text.replace("\n", ",").split(","):
-        cleaned = candidate.strip()
-        if cleaned:
-            yield cleaned
-
-
-_DEFAULT_SPEEDTEST_ENTITY_IDS: tuple[str, ...] = tuple(
-    dict.fromkeys(
-        candidate
-        for raw in DEFAULT_SPEEDTEST_ENTITIES
-        for candidate in _split_entity_candidates(str(raw))
-    )
-) or (
-    "sensor.speedtest_download",
-    "sensor.speedtest_upload",
-    "sensor.speedtest_ping",
-)
-
-
-def _normalize_speedtest_entity_ids(raw: Any) -> list[str]:
-    """Normalize speedtest entity identifiers from options/data into a stable list."""
-
-    normalized: dict[str, None] = {}
-
-    def _add_from_text(text: str) -> None:
-        for candidate in _split_entity_candidates(text):
-            if candidate not in normalized:
-                normalized[candidate] = None
-
-    if isinstance(raw, str):
-        _add_from_text(raw)
-    elif isinstance(raw, (list, tuple, set)):
-        for candidate in raw:
-            if isinstance(candidate, str):
-                _add_from_text(candidate)
-            elif candidate is not None:
-                text = str(candidate).strip()
-                if text:
-                    if text not in normalized:
-                        normalized[text] = None
-
-    if not normalized:
-        return list(_DEFAULT_SPEEDTEST_ENTITY_IDS)
-
-    return list(normalized)
-
-
-def _resolve_speedtest_interval_seconds(
-    options: Mapping[str, Any], data: Mapping[str, Any]
-) -> int:
-    """Determine the configured speedtest interval in seconds with legacy support."""
-
-    def _coerce_int(value: Any) -> int | None:
-        try:
-            if value is None:
-                return None
-            return int(value)
-        except (TypeError, ValueError):
-            return None
-
-    speedtest_interval_seconds: int | None = None
-    for source in (options, data):
-        candidate = _coerce_int(source.get(CONF_SPEEDTEST_INTERVAL))
-        if candidate is not None:
-            speedtest_interval_seconds = candidate
-            break
-
-    legacy_minutes: int | None = None
-    for source in (options, data):
-        candidate = _coerce_int(source.get(LEGACY_CONF_SPEEDTEST_INTERVAL_MIN))
-        if candidate is not None:
-            legacy_minutes = candidate
-            break
-
-    if legacy_minutes is not None:
-        legacy_seconds = max(0, legacy_minutes) * 60
-        if (
-            legacy_minutes != DEFAULT_SPEEDTEST_INTERVAL_MINUTES
-            or speedtest_interval_seconds is None
-        ):
-            speedtest_interval_seconds = legacy_seconds
-
-    if speedtest_interval_seconds is None:
-        speedtest_interval_seconds = DEFAULT_SPEEDTEST_INTERVAL
-
-    return max(0, speedtest_interval_seconds)
-
-
-def _normalize_wifi_option(value: Any) -> Optional[str]:
-    if value in (None, ""):
-        return None
-    if isinstance(value, bool):
-        return None
-    if isinstance(value, (int, float)):
-        cleaned = str(value).strip()
-        return cleaned or None
-    if isinstance(value, str):
-        cleaned = value.strip()
-        return cleaned or None
-    cleaned = str(value).strip()
-    return cleaned or None
-
 
 async def async_setup(hass: "HomeAssistant", config: "ConfigType") -> bool:
     """Set up the UniFi Gateway Dashboard Analyzer component."""
@@ -163,52 +41,11 @@ async def async_setup_entry(hass: "HomeAssistant", entry: "ConfigEntry") -> bool
     """Set up UniFi Gateway Dashboard Analyzer from a config entry."""
     _LOGGER.debug("Setting up config entry %s", entry.title)
 
-    from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
-
-    client_kwargs: dict[str, Any] = {
-        "host": entry.data[CONF_HOST],
-        "port": entry.data.get(CONF_PORT, DEFAULT_PORT),
-        "site_id": entry.data.get(CONF_SITE_ID, DEFAULT_SITE),
-        "ssl_verify": entry.data.get(CONF_VERIFY_SSL, DEFAULT_VERIFY_SSL),
-        "use_proxy_prefix": entry.data.get(
-            CONF_USE_PROXY_PREFIX, DEFAULT_USE_PROXY_PREFIX
-        ),
-        "timeout": entry.data.get(CONF_TIMEOUT, DEFAULT_TIMEOUT),
-    }
-
-    username = entry.data.get(CONF_USERNAME)
-    password = entry.data.get(CONF_PASSWORD)
-    if username:
-        client_kwargs["username"] = username
-    if password:
-        client_kwargs["password"] = password
-
-    client_kwargs["instance_hint"] = entry.entry_id  # ensures stable unique_id across restarts
-
-    client_factory = partial(UniFiOSClient, **client_kwargs)
-
+    client = await async_get_client(hass, entry)
     options = entry.options or {}
 
-    speedtest_interval_seconds = _resolve_speedtest_interval_seconds(options, entry.data)
-
-    wifi_guest = _normalize_wifi_option(options.get(CONF_WIFI_GUEST))
-    if wifi_guest is None:
-        wifi_guest = _normalize_wifi_option(entry.data.get(CONF_WIFI_GUEST))
-    wifi_iot = _normalize_wifi_option(options.get(CONF_WIFI_IOT))
-    if wifi_iot is None:
-        wifi_iot = _normalize_wifi_option(entry.data.get(CONF_WIFI_IOT))
-
-    try:
-        client: UniFiOSClient = await hass.async_add_executor_job(client_factory)
-    except AuthError as err:
-        _LOGGER.error("Authentication failed while setting up entry %s: %s", entry.entry_id, err)
-        raise ConfigEntryAuthFailed("Authentication with UniFi controller failed") from err
-    except ConnectivityError as err:
-        _LOGGER.error("Connectivity issue while setting up entry %s: %s", entry.entry_id, err)
-        raise ConfigEntryNotReady(f"Cannot connect to UniFi controller: {err}") from err
-    except APIError as err:
-        _LOGGER.error("Controller error while setting up entry %s: %s", entry.entry_id, err)
-        raise ConfigEntryNotReady(f"UniFi controller error: {err}") from err
+    speedtest_interval_seconds = resolve_speedtest_interval_seconds(options, entry.data)
+    wifi_guest, wifi_iot = get_wifi_settings(options, entry.data)
 
     coordinator = UniFiGatewayDataUpdateCoordinator(
         hass,
@@ -223,7 +60,7 @@ async def async_setup_entry(hass: "HomeAssistant", entry: "ConfigEntry") -> bool
         interval_minutes = max(5, round(speedtest_interval_seconds / 60))
 
     raw_entities = options.get(CONF_SPEEDTEST_ENTITIES, DEFAULT_SPEEDTEST_ENTITIES)
-    entity_ids = _normalize_speedtest_entity_ids(raw_entities)
+    entity_ids = normalize_speedtest_entity_ids(raw_entities)
 
     async def _noop_result_callback(
         *, success: bool, duration_ms: int, error: str | None, trace_id: str
@@ -271,8 +108,8 @@ async def async_setup_entry(hass: "HomeAssistant", entry: "ConfigEntry") -> bool
         entry.entry_id,
     )
 
-    await _async_migrate_speedtest_button_unique_id(hass, entry)
-    await _async_migrate_interface_unique_ids(hass, entry, client, coordinator.data)
+    await async_migrate_speedtest_button_unique_id(hass, entry)
+    await async_migrate_interface_unique_ids(hass, entry, client, coordinator.data)
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
 
@@ -315,120 +152,5 @@ async def async_unload_entry(hass: "HomeAssistant", entry: "ConfigEntry") -> boo
             "UniFi Gateway Dashboard Analyzer entry %s unloaded", entry.entry_id
         )
     return unload_ok
-
-
-async def _async_migrate_speedtest_button_unique_id(
-    hass: "HomeAssistant", entry: "ConfigEntry"
-) -> None:
-    """Ensure the Run Speedtest button unique ID is namespaced per config entry."""
-
-    from homeassistant.helpers import entity_registry as er
-
-    try:
-        from .utils import build_speedtest_button_unique_id
-    except ImportError:  # pragma: no cover - defensive guard
-        return
-
-    old_unique_id = "unifi_gateway_refactored_run_speedtest"
-    new_unique_id = build_speedtest_button_unique_id(entry.entry_id)
-
-    if old_unique_id == new_unique_id:
-        return
-
-    migrated = False
-
-    async def _migrate(entity_entry: er.RegistryEntry) -> dict[str, str] | None:
-        nonlocal migrated
-        if entity_entry.config_entry_id != entry.entry_id:
-            return None
-        if entity_entry.unique_id != old_unique_id:
-            return None
-        migrated = True
-        return {"new_unique_id": new_unique_id}
-
-    await er.async_migrate_entries(hass, DOMAIN, _migrate)
-
-    if migrated:
-        _LOGGER.info(
-            "Migrated Run Speedtest button unique ID for entry %s", entry.entry_id
-        )
-
-
-async def _async_migrate_interface_unique_ids(
-    hass: "HomeAssistant",
-    entry: "ConfigEntry",
-    client: UniFiOSClient,
-    data: UniFiGatewayData | None,
-) -> None:
-    """Normalize WAN/LAN/WLAN sensor unique IDs."""
-
-    from homeassistant.helpers import entity_registry as er
-    from .sensor import (
-        _wan_identifier_candidates,
-        build_lan_unique_id,
-        build_wan_unique_id,
-        build_wlan_unique_id,
-    )
-
-    if not data:
-        return
-
-    mapping: dict[str, str] = {}
-    instance_prefix = f"unifigw_{client.instance_key()}"
-
-    for link in data.wan_links:
-        if not isinstance(link, dict):
-            continue
-        link_id = str(link.get("id") or link.get("_id") or link.get("ifname") or "wan")
-        link_name = link.get("name") or link_id
-        identifiers = _wan_identifier_candidates(link_id, link_name, link)
-        canonical = (sorted(identifiers) or [link_id])[0]
-        old_key = hashlib.sha256(canonical.encode()).hexdigest()[:12]
-        for suffix in ("status", "ip", "ipv6", "isp"):
-            old_uid = f"{instance_prefix}_wan_{old_key}_{suffix}"
-            new_uid = build_wan_unique_id(entry.entry_id, link, suffix)
-            mapping[old_uid] = new_uid
-
-    for network in data.lan_networks:
-        if not isinstance(network, dict):
-            continue
-        net_id = str(
-            network.get("_id") or network.get("id") or network.get("name") or "lan"
-        )
-        old_uid = f"{instance_prefix}_lan_{net_id}_clients"
-        new_uid = build_lan_unique_id(entry.entry_id, network)
-        mapping[old_uid] = new_uid
-
-    for wlan in data.wlans:
-        if not isinstance(wlan, dict):
-            continue
-        ssid = wlan.get("name") or wlan.get("ssid") or wlan.get("_id") or wlan.get("id")
-        if not ssid:
-            continue
-        old_uid = f"{instance_prefix}_wlan_{ssid}_clients"
-        new_uid = build_wlan_unique_id(entry.entry_id, wlan)
-        mapping[old_uid] = new_uid
-
-    if not mapping:
-        _LOGGER.debug("No interface unique ID migrations required for entry %s", entry.entry_id)
-        return
-
-    async def _migrate(entity_entry: er.RegistryEntry) -> dict[str, str] | None:
-        if entity_entry.config_entry_id != entry.entry_id:
-            return None
-        unique_id = entity_entry.unique_id
-        if unique_id is None:
-            return None
-        new_uid = mapping.get(unique_id)
-        if new_uid:
-            return {"new_unique_id": new_uid}
-        return None
-
-    await er.async_migrate_entries(hass, DOMAIN, _migrate)
-    _LOGGER.info(
-        "Migrated %s interface entities to normalized unique IDs for entry %s",
-        len(mapping),
-        entry.entry_id,
-    )
