+++ tests/stubs/homeassistant/const.py
@@ -21,4 +21,9 @@ class Platform(str, Enum):
     SENSOR = "sensor"
 
 
-__all__ = ["Platform", "UnitOfTime", "TIME_MILLISECONDS"]
+CONF_HOST = "host"
+CONF_PORT = "port"
+CONF_USERNAME = "username"
+CONF_PASSWORD = "password"
+
+__all__ = ["Platform", "UnitOfTime", "TIME_MILLISECONDS", "CONF_HOST", "CONF_PORT", "CONF_USERNAME", "CONF_PASSWORD"]
